- move ChartPrice, ChartStake, EconomicsCard to widgets - cleanup duplicate .tooltip css
