+++ src/widgets/ChartPrice/ChartPrice.tsx
