+++ src/widgets/ChartPrice/index.ts
