+++ src/widgets/ChartPrice/styles.module.scss
@@ -4,19 +4,18 @@
   margin: 1rem 0;
   padding: 1rem 0;
   display: flex;
+  flex-direction: column;
   min-height: 18.75rem;
   position: relative;
-  flex-direction: column;
 
   @media (min-width: 768px) {
-    padding: 1.5rem 0;
+    padding: 2rem 0;
     margin: 2rem 0;
   }
 
   @media (min-width: 1200px) {
-    width: 41.85%;
+    width: 37.5%;
     margin: 0;
-    margin-left: 1rem;
     margin-bottom: 1rem;
   }
 
@@ -27,8 +26,8 @@
     bottom: 0;
     width: 100%;
     height: 100%;
-    display: flex;
     align-items: center;
+    display: flex;
   }
 
   .wrapper {
@@ -37,7 +36,6 @@
     justify-content: space-between;
     color: var(--primary);
     padding: 0 1rem 1rem;
-    pointer-events: none;
     position: relative;
     z-index: 2;
     pointer-events: none;
@@ -66,20 +64,26 @@
     @media (min-width: 992px) {
       font-size: 2rem;
     }
-
-    span {
-      color: var(--primary-100);
-      font-size: 1.5rem;
-    }
   }
 
   .change {
     display: flex;
     align-items: center;
 
+    &.up {
+      color: var(--primary-200);
+    }
+
+    &.down {
+      color: var(--red);
+    }
+
+    &.neutral {
+      color: var(--neutral-300);
+    }
+
     .percentage {
       font-size: 0.75rem;
-      color: var(--primary-200);
       margin-left: 0.33rem;
     }
   }
@@ -88,8 +92,8 @@
     display: flex;
     align-items: flex-start;
     justify-content: flex-start;
-    flex-wrap: wrap;
     padding: 3rem 1rem 0;
+    flex-wrap: wrap;
     margin-top: auto;
 
     @media (min-width: 768px) {
@@ -97,8 +101,8 @@
     }
 
     .statistic {
-      margin-top: 1rem;
       margin-right: 2rem;
+      margin-top: 1rem;
 
       &:last-child {
         margin-right: 0;
@@ -116,46 +120,5 @@
         }
       }
     }
-  }
-}
-
-.tooltip {
-  border: 2px solid #24f7dd33;
-  border-radius: 1rem;
-  padding: 0.625rem;
-  padding-left: 1.75rem;
-  line-height: 1;
-  color: var(--primary);
-  font-family: 'Roobert SemiBold';
-  font-weight: 600;
-  position: relative;
-  overflow: hidden;
-
-  &:before {
-    content: '';
-    left: 0.625rem;
-    width: 0.625rem;
-    top: 50%;
-    -webkit-transform: translateY(-50%);
-    transform: translateY(-50%);
-    position: absolute;
-    border-radius: 50%;
-    background-color: var(--primary);
-    height: 0.625rem;
-  }
-
-  &:after {
-    content: '';
-    left: 0;
-    right: 0;
-    top: 0;
-    bottom: 0;
-    position: absolute;
-    background: linear-gradient(
-      to bottom,
-      black 0%,
-      rgba(var(--primary-rgb), 0.15) 100%
-    );
-    z-index: -1;
   }
 }
