+++ src/widgets/ChartStake/ChartStake.tsx
