+++ src/widgets/ChartStake/index.ts
