+++ src/widgets/EconomicsCard/EconomicsCard.tsx
