+++ src/widgets/EconomicsCard/index.ts
