+++ src/widgets/EconomicsCard/styles.module.scss
