+++ src/widgets/index.ts
@@ -2,6 +2,9 @@ export * from './AccountsStatsCard';
 export * from './BlockHeightStatsCard';
 export * from './BlockProgressRing';
 export * from './ChartContractsTransactions';
+export * from './ChartPrice';
+export * from './ChartStake';
+export * from './EconomicsCard';
 export * from './EpochProgressRing';
 export * from './HeroApplications';
 export * from './HeroHome';
